# src/docs/swagger.yaml
definitions:
  dto.CarModelColorResponse:
    properties:
      color:
        $ref: '#/definitions/dto.ColorResponse'
      id:
        type: integer
    type: object
  dto.CarModelResponse:
    properties:
      carModelColors:
        items:
          $ref: '#/definitions/dto.CarModelColorResponse'
        type: array
      carType:
        $ref: '#/definitions/dto.CarTypeResponse'
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      gearbox:
        $ref: '#/definitions/dto.GearBoxResponse'
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CarTypeResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CityResponse:
    properties:
      country:
        $ref: '#/definitions/dto.CountryResponse'
      id:
        type: integer
      name:
        type: string
    type: object
  dto.ColorResponse:
    properties:
      hex:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CompanyResponse:
    properties:
      country:
        $ref: '#/definitions/dto.CountryResponse'
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CountryResponse:
    properties:
      cities:
        items:
          $ref: '#/definitions/dto.CityResponse'
        type: array
      companies:
        items:
          $ref: '#/definitions/dto.CompanyResponse'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CreateCarModelColorRequest:
    properties:
      carModelId:
        type: integer
      colorId:
        type: integer
    required:
    - carModelId
    - colorId
    type: object
  dto.CreateCarModelRequest:
    properties:
      carTypeId:
        type: integer
      companyId:
        type: integer
      gearboxId:
        type: integer
      name:
        maxLength: 15
        type: string
    required:
    - carTypeId
    - companyId
    - gearboxId
    - name
    type: object
...
  dto.CreateCompanyRequest:
    properties:
      countryId:
        type: integer
      name:
        maxLength: 15
        type: string
    required:
    - countryId
    type: object
...
  dto.Filter:
    properties:
      filter_type:
        type: string
      from:
        type: string
      to:
        type: string
      type:
        type: string
    type: object
  dto.GearBoxResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/dto.Filter'
        type: object
      page_number:
        type: integer
      page_size:
        type: integer
      sort:
        items:
          $ref: '#/definitions/dto.SortDTO'
        type: array
    type: object
...
  dto.SortDTO:
    properties:
      col_id:
        type: string
      sort:
        type: string
    type: object
  dto.UpdateCarModelColorRequest:
    properties:
      carModelId:
        type: integer
      colorId:
        type: integer
    type: object
  dto.UpdateCarModelRequest:
    properties:
      carTypeId:
        type: integer
      companyId:
        type: integer
      gearboxId:
        type: integer
      name:
        maxLength: 15
        type: string
    required:
    - name
    type: object
  dto.UpdateCarTypeRequest:
    properties:
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  helper.Response:
    properties:
      error: {}
      result: {}
      result_code:
        type: integer
      success:
        type: boolean
      validation_error:
        items:
          $ref: '#/definitions/validators.ValidationError'
        type: array
    type: object
  validators.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/car-color/create:
    post:
      consumes:
      - application/json
      description: Create a CarModelColor
      parameters:
      - description: Create a CarModelColor
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCarModelColorRequest'
      responses:
        "201":
          description: CarModelColor response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CarModelColorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a CarModelColor
      tags:
      - CarModelColor
  /v1/car-color/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Get a CarModelColor
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: CarModelColor response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a CarModelColor
      tags:
      - CarModelColor
  /v1/car-color/filter:
    post:
      consumes:
      - application/json
      description: Get CarModelColor
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.PaginationInputWithFilter'
      responses:
        "200":
          description: City response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get CarModelColor
      tags:
      - CarModelColor
  /v1/car-color/get/{id}:
    get:
      consumes:
      - application/json
      description: Delete a CarModelColor
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a CarModelColor
      tags:
      - CarModelColor
  /v1/car-color/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a CarModelColor
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a CarModelColor
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCarModelColorRequest'
      responses:
        "200":
          description: CarModelColor response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CarModelColorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a CarModelColor
      tags:
      - CarModelColor
  /v1/car-model/create:
    post:
      consumes:
      - application/json
      description: Create a CarModel
      parameters:
      - description: Create a CarModel
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCarModelRequest'
      responses:
        "201":
          description: CarModel response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CarModelResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a CarModel
      tags:
      - CarModel
  /v1/car-model/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Get a CarModel
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: CarModel response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a CarModel
      tags:
      - CarModel
  /v1/car-model/filter:
    post:
      consumes:
      - application/json
      description: Get CarModel
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.PaginationInputWithFilter'
      responses:
        "200":
          description: City response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get CarModel
      tags:
      - CarModel
  /v1/car-model/get/{id}:
    get:
      consumes:
      - application/json
      description: Delete a CarModel
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a CarModel
      tags:
      - CarModel
  /v1/car-model/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a CarModel
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a CarModel
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCarModelRequest'
      responses:
        "200":
          description: CarModel response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CarModelResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a CarModel
      tags:
      - CarModel
  /v1/car-type/create:
    post:
      consumes:
      - application/json
      description: Create a Company
      parameters:
      - description: Create a Company
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCompanyRequest'
      responses:
        "201":
          description: Company response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CompanyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a Company
      tags:
      - Company
  /v1/car-type/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Get a CarType
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: CarType response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a CarType
      tags:
      - CarType
  /v1/car-type/filter:
    post:
      consumes:
      - application/json
      description: Get CarType
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.PaginationInputWithFilter'
      responses:
        "200":
          description: City response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get CarType
      tags:
      - CarType
  /v1/car-type/get/{id}:
    get:
      consumes:
      - application/json
      description: Delete a CarType
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a CarType
      tags:
      - CarType
  /v1/car-type/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a CarType
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a CarType
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCarTypeRequest'
      responses:
        "200":
          description: CarType response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CarTypeResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a CarType
      tags:
      - CarType
...
swagger: "2.0"